requirements.txt
# Core ML/AI Libraries
torch>=2.1.0
transformers>=4.36.0
trl>=0.7.0
vllm>=0.2.7
accelerate>=0.24.0
bitsandbytes>=0.41.0

# PEFT for LoRA and other parameter-efficient fine-tuning
peft>=0.7.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
starlette>=0.27.0

# Async and Concurrency
asyncio-throttle>=1.0.2
aiofiles>=23.2.1
httpx>=0.25.0

# Streaming and Messaging
kafka-python>=2.0.2
aiokafka>=0.9.0

# Cloud Storage
boto3>=1.34.0
botocore>=1.34.0

# Configuration Management
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0

# Neural Network Utilities
einops>=0.7.0
safetensors>=0.4.0
bert-score>=0.3.13

# CLI and Development
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# Testing and Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database (optional, for local persistence)
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0

# Utilities
tqdm>=4.66.0
psutil>=5.9.0
schedule>=1.2.0
backoff>=2.2.0

# JSON and Serialization
orjson>=3.9.0
msgpack>=1.0.7

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0

# Environment Management
python-multipart>=0.0.6
email-validator>=2.1.0

# Mathematical Operations

# GPU Monitoring (optional, not directly integrated yet)

# Development Dependencies (optional)
jupyter>=1.0.0
ipython>=8.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Production Dependencies (uvicorn is primary for now)

# Linting and Formatting
pre-commit>=3.6.0
isort>=5.12.0
autoflake>=2.2.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Profiling
py-spy>=0.3.14
memory-profiler>=0.61.0

# System Dependencies (may require system installation)
# For tokenizers and fast operations
tokenizers>=0.15.0

# For efficient tensor operations (optional, advanced)

# For model quantization (optional, advanced)

# Version pins for stability
certifi>=2023.11.17
charset-normalizer>=3.3.2
idna>=3.6
urllib3>=2.1.0 